requirements.txt
django
django-rest-framework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
django-redis
dotenv
psycopg2-binary